# Conflicts:
#	library/leon/collection/List.scala
#	src/main/scala/leon/grammars/ExpressionGrammar.scala
#	src/main/scala/leon/synthesis/rules/CEGIS.scala
#	src/main/scala/leon/synthesis/rules/CEGISLike.scala
#	src/main/scala/leon/synthesis/rules/CEGLESS.scala
#	src/main/scala/leon/synthesis/rules/StringRender.scala
#	src/main/scala/leon/synthesis/rules/unused/SygusCVC4.scala
#	src/main/scala/leon/synthesis/rules/unused/TEGLESS.scala

+++ src/test/resources/regression/synthesis/Holes/Hole1.scala
@@ -1,11 +1,11 @@
-/* Copyright 2009-2015 EPFL, Lausanne */
+/* Copyright 2009-2016 EPFL, Lausanne */
 
 import leon.lang._
 import leon.collection._
 import leon.lang.synthesis._
 
 object Holes {
-  def abs3(a: Int) = {
+  def abs3(a: BigInt) = {
     if (?(a > 0, a < 0)) {
       a
     } else {
